# openapi.yaml
openapi: 3.0.1
info:
  title: Tournaments API
  version: 1.0.0
  description: >-
    Una API con la información de los torneos que se organizan fuera de la liga
    regular de un eSport
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor local
paths:
  /register:
    summary: Registro de un usuario
    description: >-
      Solo es necesario registrarse para poder realizar operaciones de borrado
      en las tablas
    post:
      summary: Registrar nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDto'
            examples:
              example1:
                summary: Registro de usuario válido
                value:
                  username: visitor
                  pasword: 1234
              example2:
                summary: Registro de usuario no válido
                value:
                  username: visitor
      responses:
        '201':
          description: Exito al registrar al usuario
          content:
            application/json:
              schema:
                type: string
              examples:
                example1:
                  summary: Usuario creado
                  value: visitor created
        '400':
          description: Los datos de registro están incompletos o no son correctos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                example2:
                  summary: Datos incompletos/no válidos
                  value:
                    status: 400
                    message: Bad request
                    errorMessages:
                      password: Password required
        '409':
          description: El usuario ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                example2:
                  summary: Usuario ya existe
                  value:
                    status: 409
                    message: User already exist
                    errorMessages: null
      tags:
        - Auth
  /login:
    summary: >-
      Iniciar sesion con uusario y contraseña y obtener el token de
      autenticación
    description: El token solo es necesario para las operaciones de borrado
    post:
      summary: Identificarse con usuario y contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDto'
            examples:
              example1:
                summary: Login de usuario válido
                value:
                  username: visitor
                  pasword: 1234
              example2:
                summary: Login con password incorrecto
                value:
                  username: visitor
                  password: 3897
      responses:
        '200':
          description: Exito al iniciar sesión, devuelve el token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenDto'
              examples:
                example1:
                  summary: Token generado
                  value:
                    token: eyJhbGciOiJIUzI1NiJ9..
        '401':
          description: Contraseña incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                example2:
                  summary: Contraseña incorrecta
                  value:
                    code: 401
                    message: Invalid credentials
                    errorMessage: null
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                example3:
                  summary: Usuario no encontrado
                  value:
                    code: 404
                    message: User not found
                    errorMessage: null
      tags:
        - Auth
...
components:
  schemas:
    UserLoginDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    AuthTokenDto:
      type: object
      properties:
        token:
          type: string
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...